+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/timeline/elements/AudioMessageViewModel.kt
@@ -1,15 +1,20 @@
 package de.connect2x.trixnity.messenger.viewmodel.room.timeline.elements
 
+import de.connect2x.trixnity.messenger.util.FileTransferProgressElement
 import de.connect2x.trixnity.messenger.viewmodel.MatrixClientViewModelContext
 import de.connect2x.trixnity.messenger.viewmodel.UserInfoElement
 import de.connect2x.trixnity.messenger.viewmodel.room.timeline.OpenModalCallback
+import de.connect2x.trixnity.messenger.viewmodel.room.timeline.elements.util.Thumbnails
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.stateIn
 import net.folivo.trixnity.client.store.TimelineEvent
+import net.folivo.trixnity.clientserverapi.model.media.FileTransferProgress
 import net.folivo.trixnity.core.model.UserId
 import net.folivo.trixnity.core.model.events.m.room.RoomMessageEventContent
+import org.koin.core.component.get
 
 
 interface AudioMessageViewModelFactory {
@@ -27,6 +32,7 @@ interface AudioMessageViewModelFactory {
         sender: Flow<UserInfoElement>,
         invitation: Flow<String?>,
         onOpenModal: OpenModalCallback,
+        mediaUploadProgress: MutableStateFlow<FileTransferProgress?>
     ): AudioMessageViewModel {
         return AudioMessageViewModelImpl(
             viewModelContext,
@@ -42,6 +48,7 @@ interface AudioMessageViewModelFactory {
             sender,
             invitation,
             onOpenModal,
+            mediaUploadProgress
         )
     }
 
@@ -67,6 +74,7 @@ open class AudioMessageViewModelImpl(
     sender: Flow<UserInfoElement>,
     invitation: Flow<String?>,
     private val onOpenModal: OpenModalCallback,
+    mediaUploadProgress: MutableStateFlow<FileTransferProgress?>
 ) : AudioMessageViewModel, AbstractFileBasedMessageViewModel(viewModelContext, content),
     MatrixClientViewModelContext by viewModelContext {
     override val duration: Int? = content.info?.duration
@@ -76,6 +84,9 @@ open class AudioMessageViewModelImpl(
         sender.stateIn(coroutineScope, SharingStarted.WhileSubscribed(), UserInfoElement("", UserId("")))
     override val showSender: StateFlow<Boolean> =
         showSender.stateIn(coroutineScope, SharingStarted.WhileSubscribed(), true)
+    private val thumbnails = get<Thumbnails>()
+    override val uploadProgress: StateFlow<FileTransferProgressElement?> = thumbnails.mapProgressToProgressElement(mediaUploadProgress)
+        .stateIn(coroutineScope, SharingStarted.WhileSubscribed(), null)
 
     override fun openAudio() {
         // TODO if you have audio working, replace with: 'url?.let { onOpenModal(OpenModalType.AUDIO, it, encryptedFile) }'