+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/timeline/elements/FileBasedMessageViewModel.kt
@@ -27,6 +27,7 @@ interface FileBasedMessageViewModel : RoomMessageViewModel {
     val saveFileDialogOpen: StateFlow<Boolean>
     val downloadProgress: StateFlow<FileTransferProgressElement?>
     val downloadSuccessful: StateFlow<Boolean?>
+    val uploadProgress: StateFlow<FileTransferProgressElement?>
     val fileName: String
     val fileSize: Int?
     val fileMimeType: String?